refactor(activemodel,activerecord): converge serialization test layout, respond_to?, create_table raise, inverse_of (RFC 0115, RFC 0051)
- Move the 31 TS-only `it(...)`s out of the mirrored
  activemodel/src/serialization.test.ts into a serialization.trails.test.ts
  sibling. serialization_test.rb goes Extra 31 -> 0, OK 20 unchanged;
  activemodel matched total unchanged at 963/963.
- Drop respond_to?'s dead middle arm (attribute_methods.rb:531-535): a JS
  receiver has no string-named private methods, so both `super` sends always
  agree and the arm can never be taken. isRespondToWithoutAttributes no longer
  accepts the parameter it discarded.
- create_table raises ArgumentError, not a bare Error, and guards on key
  presence like Rails' `options.key?(:if_not_exists)`
  (schema_statements.rb:297-299).
- Fold _dispatchInvert back into inverseOf and raise Rails' four-line
  IrreversibleMigration message (command_recorder.rb:114-123).
